64bit & Lion Full Screen support

+++ ASIDataCompressor.m
@@ -78,7 +78,7 @@
 	int status;
 	
 	zStream.next_in = bytes;
-	zStream.avail_in = (unsigned int)length;
+	zStream.avail_in = (NSUInteger )length;
 	zStream.avail_out = 0;
 
 	NSInteger bytesProcessedAlready = zStream.total_out;
@@ -89,7 +89,7 @@
 		}
 		
 		zStream.next_out = [outputData mutableBytes] + zStream.total_out-bytesProcessedAlready;
-		zStream.avail_out = (unsigned int)([outputData length] - (zStream.total_out-bytesProcessedAlready));
+		zStream.avail_out = (NSUInteger )([outputData length] - (zStream.total_out-bytesProcessedAlready));
 		status = deflate(&zStream, shouldFinish ? Z_FINISH : Z_NO_FLUSH);
 		
 		if (status == Z_STREAM_END) {
