+++ ArticleListView.m
@@ -67,8 +67,8 @@
 	-(void)endMainFrameLoad;
 @end
 
-static const int MA_Minimum_ArticleList_Pane_Width = 80;
-static const int MA_Minimum_Article_Pane_Width = 80;
+static const CGFloat MA_Minimum_ArticleList_Pane_Width = 80;
+static const CGFloat MA_Minimum_Article_Pane_Width = 80;
 
 @implementation ArticleListView
 
@@ -167,7 +167,7 @@ static const int MA_Minimum_Article_Pane_Width = 80;
  * Make sure the article pane width isn't shrunk beyond a minimum width. Otherwise it looks
  * untidy.
  */
--(float)splitView:(NSSplitView *)sender constrainMinCoordinate:(float)proposedMin ofSubviewAt:(int)offset
+-(CGFloat)splitView:(NSSplitView *)sender constrainMinCoordinate:(CGFloat)proposedMin ofSubviewAt:(NSInteger)offset
 {
 	return (sender == splitView2 && offset == 0) ? MA_Minimum_ArticleList_Pane_Width : proposedMin;
 }
@@ -176,7 +176,7 @@ static const int MA_Minimum_Article_Pane_Width = 80;
  * Make sure that the article pane isn't shrunk beyond a minimum size otherwise the splitview
  * or controls within it start resizing odd.
  */
--(float)splitView:(NSSplitView *)sender constrainMaxCoordinate:(float)proposedMax ofSubviewAt:(int)offset
+-(CGFloat)splitView:(NSSplitView *)sender constrainMaxCoordinate:(CGFloat)proposedMax ofSubviewAt:(NSInteger)offset
 {
 	if (sender == splitView2 && offset == 0)
 	{
@@ -193,7 +193,7 @@ static const int MA_Minimum_Article_Pane_Width = 80;
  */
 -(void)splitView:(NSSplitView *)sender resizeSubviewsWithOldSize:(NSSize)oldSize
 {
-	float dividerThickness = [sender dividerThickness];
+	CGFloat dividerThickness = [sender dividerThickness];
 	id sv1 = [[sender subviews] objectAtIndex:0];
 	id sv2 = [[sender subviews] objectAtIndex:1];
 	NSRect leftFrame = [sv1 frame];
@@ -252,7 +252,7 @@ static const int MA_Minimum_Article_Pane_Width = 80;
  * Called from the webview when the user positions the mouse over an element. If it's a link
  * then echo the URL to the status bar like Safari does.
  */
--(void)webView:(WebView *)sender mouseDidMoveOverElement:(NSDictionary *)elementInformation modifierFlags:(unsigned int)modifierFlags
+-(void)webView:(WebView *)sender mouseDidMoveOverElement:(NSDictionary *)elementInformation modifierFlags:(NSUInteger )modifierFlags
 {
 	NSURL * url = [elementInformation valueForKey:@"WebElementLinkURL"];
 	[controller setStatusMessage:(url ? [url absoluteString] : @"") persist:NO];
@@ -320,7 +320,7 @@ static const int MA_Minimum_Article_Pane_Width = 80;
 	NSArray * dataArray = [prefs arrayForKey:MAPref_ArticleListColumns];
 	Database * db = [Database sharedDatabase];
 	Field * field;
-	unsigned int index;
+	NSUInteger  index;
 	
 	for (index = 0; index < [dataArray count];)
 	{
@@ -796,7 +796,7 @@ static const int MA_Minimum_Article_Pane_Width = 80;
  * Support special key codes. If we handle the key, return YES otherwise
  * return NO to allow the framework to pass it on for default processing.
  */
--(BOOL)handleKeyDown:(unichar)keyChar withFlags:(unsigned int)flags
+-(BOOL)handleKeyDown:(unichar)keyChar withFlags:(NSUInteger )flags
 {
 	return [controller handleKeyDown:keyChar withFlags:flags];
 }
@@ -1811,7 +1811,7 @@ static const int MA_Minimum_Article_Pane_Width = 80;
 	if ([articleList numberOfSelectedRows] > 0)
 	{
 		NSIndexSet * rowIndexes = [articleList selectedRowIndexes];
-		unsigned int rowIndex = [rowIndexes firstIndex];
+		NSUInteger  rowIndex = [rowIndexes firstIndex];
 
 		articleArray = [NSMutableArray arrayWithCapacity:[rowIndexes count]];
 		while (rowIndex != NSNotFound)
